spline #108 addressing PR comments: move Processing traits to a single file. Add scala docs.
